Rethinking concurrent approach yet again
I think I have a better understanding now of how the concurrency is
working; specifically, what's blocking (knock on wood). The whole setup,
for the most part, was blocking on reading stdin, which is fine, and
what I wanted, but not for an async `PY_TARGET`.

The new `worker` takes a more principled approach to managing blocking,
isolating to the `get` call to the synchronous queue `inputs`. I think
running it with a separate executor will make it properly awaitable,
allowing an async `PY_TARGET` to take some CPU while _main waits on the
queue.